requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPDF2==3.0.1
requests==2.32.3
schedule==1.2.2
soupsieve==2.6
SQLite4==0.1.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
